Rearranged one-time use schemas
